Split the render into render and review data
