+++ openpype/hosts/fusion/plugins/publish/extract_render_local.py
@@ -4,16 +4,12 @@ from openpype.pipeline import publish
 from openpype.hosts.fusion.api import comp_lock_and_undo_chunk
 
 
-class Fusionlocal(pyblish.api.InstancePlugin,
-                  publish.ColormanagedPyblishPluginMixin):
-    """Render the current Fusion composition locally.
+class FusionRenderLocal(
+    pyblish.api.InstancePlugin, publish.ColormanagedPyblishPluginMixin
+):
+    """Render the current Fusion composition locally."""
 
-    Extract the result of savers by starting a comp render
-    This will run the local render of Fusion.
-
-    """
-
-    order = pyblish.api.ExtractorOrder - 0.1
+    order = pyblish.api.ExtractorOrder - 0.2
     label = "Render Local"
     hosts = ["fusion"]
     families = ["render.local"]
@@ -32,38 +28,6 @@ class Fusionlocal(pyblish.api.InstancePlugin,
                 tsk=instance.data["task"],
             )
         )
-
-        frame_start = context.data["frameStartHandle"]
-        frame_end = context.data["frameEndHandle"]
-        path = instance.data["path"]
-        output_dir = instance.data["outputDir"]
-
-        basename = os.path.basename(path)
-        head, ext = os.path.splitext(basename)
-        files = [
-            f"{head}{str(frame).zfill(4)}{ext}"
-            for frame in range(frame_start, frame_end + 1)
-        ]
-        repre = {
-            "name": ext[1:],
-            "ext": ext[1:],
-            "frameStart": f"%0{len(str(frame_end))}d" % frame_start,
-            "files": files,
-            "stagingDir": output_dir,
-        }
-
-        self.set_representation_colorspace(
-            representation=repre,
-            context=context,
-        )
-
-        if "representations" not in instance.data:
-            instance.data["representations"] = []
-        instance.data["representations"].append(repre)
-
-        # review representation
-        if instance.data.get("review", False):
-            repre["tags"] = ["review", "ftrackreview"]
 
     def render_once(self, context):
         """Render context comp only once, even with more render instances"""
